refactor(tools/file_tools): extract path resolution to _path_resolution
`file_tools.py` carried ~290 LOC of path-handling and post-write
helpers in front of the five SyncTool classes. Moved to a sibling
module `_path_resolution.py`:

- `LocalFileAccessViolation` (ValidationError subclass)
- `ResolvedPath` (dataclass returned by the resolver)
- `_DANGEROUS_PATH_CHARS`, `_WINDOWS_DRIVE_RE` (compiled regex constants)
- `_looks_like_outside_project(path)` (heuristic for error hints)
- `_resolve_absolute_or_strip(...)` (handles `/`-leading paths)
- `_resolve_project_path(path, cwd)` (main entry — normalize + containment)
- `_validate_path_within_working_directory(path, cwd)` (legacy wrapper)
- `_POST_WRITE_VALIDATORS` (ext → validator template)
- `_post_write_validation_hint(path)` (lookup)
- `_apply_post_write_hint(path, metadata, parts)` (item 6 helper)

file_tools.py now re-imports these from `_path_resolution` so existing
test imports (`from deile.tools.file_tools import _resolve_project_path`
etc.) keep working unchanged. The path-handling area-of-responsibility
is now testable in isolation and `file_tools.py` is focused on the
five SyncTool classes (Read/Write/Edit/List/Delete).

file_tools.py: 1791 → 1501 LOC. New file _path_resolution.py: 310 LOC.
Also dropped the now-unused `Tuple` import from file_tools.py.
